+++ clsify/blast.py
@@ -16,7 +16,7 @@ from logzero import logger
 
 def revcomp(seq):
     m = {"a": "t", "A": "T", "t": "a", "T": "A", "c": "g", "C": "G", "g": "c", "G": "C"}
-    return "".join(reversed(map(lambda x: m.get(x, x), seq)))
+    return "".join(reversed(list(map(lambda x: m.get(x, x), seq))))
 
 
 @attr.s(auto_attribs=True, frozen=True)
@@ -30,19 +30,26 @@ class Alignment:
     #: query sequence
     qseq: str
 
+    def revcomp(self):
+        return Alignment(revcomp(self.hseq), "".join(reversed(self.midline)), revcomp(self.qseq))
+
     def wrapped(self, qry_start, db_start, line_length=60):
         result = []
         for offset in range(0, len(self.hseq), line_length):
             end = min(len(self.hseq), offset + line_length)
             result += [
                 ("Sbjct %%4d %%-%ds %%s" % line_length)
-                % (qry_start + 1, self.hseq[offset:end], qry_start + line_length),
+                % (db_start + offset + 1, self.hseq[offset:end], db_start + offset + line_length),
                 ("      %%4s %%-%ds" % line_length) % ("", self.midline[offset:end]),
                 ("Query %%4d %%-%ds %%s" % line_length)
-                % (db_start + 1, self.qseq[offset:end], db_start + line_length),
+                % (qry_start + offset + 1, self.qseq[offset:end], qry_start + offset + line_length),
                 "",
             ]
         return "\n".join(result)
+
+    @staticmethod
+    def build_empty():
+        return Alignment(hseq="", midline="", qseq="")
 
 
 @attr.s(auto_attribs=True, frozen=True)
@@ -94,43 +101,21 @@ class BlastMatch:
             database_end=0,
             match_cigar="",
             match_seq="",
-            alignment=None,
+            alignment=Alignment.build_empty(),
             match_json=match_json,
         )
 
-    def to_sam_tsv(self):
-        """Return ``list`` of ``str``s with the fields."""
-        if self.is_match:
-            read_seq = self.match_seq
-            if self.database_strand != self.query_strand:
-                read_seq = revcomp(read_seq)
-            return [
-                self.query,
-                "2",
-                self.database,
-                str(self.database_start + 1),
-                "1",
-                self.match_cigar,
-                "*",
-                "0",
-                "0",
-                read_seq,
-                "*",
-            ]
-        else:
-            return [self.query, "4", "*", "0", "0", "*", "*", "0", "*", "*"]
 
-
-def _is_nucl(c):
+def is_nucl(c):
     """Helper, returns wether ``c`` is a nucleotide character."""
     return c in "acgtnACGTN"
 
 
-def _match_cigar(qseq, hseq, query_start, query_end, query_len):
+def match_cigar(qseq, hseq, query_start, query_end, query_len):
     """Return CIGAR string for the match with the given paraemters."""
     match_cigar = []
     if query_start > 0:
-        match_cigar.append([query_start + 1, "H"])
+        match_cigar.append([query_start, "H"])
     for i, (q, h) in enumerate(zip(qseq, hseq)):
         if q == "-":
             op = "D"
@@ -138,7 +123,7 @@ def _match_cigar(qseq, hseq, query_start, query_end, query_len):
             op = "I"
         else:
             op = "M"
-        if match_cigar[-1][1] == op:
+        if match_cigar and match_cigar[-1][1] == op:
             match_cigar[-1][0] += 1
         else:
             match_cigar.append([1, op])
@@ -153,28 +138,38 @@ def parse_blastn_json(database, query, match_json):
     search = match_dict["BlastOutput2"][0]["report"]["results"]["search"]
     hits = search["hits"]
     if not hits:
-        return BlastMatch.build_nomatch(query, match_json)
+        return BlastMatch.build_nomatch(os.path.basename(query), match_json)
     else:
         hsp = hits[0]["hsps"][0]
         qseq = "".join(filter(lambda x: x in "ACGTNacgtn", hsp["qseq"]))
         hseq = "".join(filter(lambda x: x in "ACGTNacgtn", hsp["hseq"]))
-        db_start = min(hsp["hit_from"], hsp["hit_to"])
-        query_start = min(hsp["query_from"], hsp["query_to"])
+        db_strand = "+" if hsp["hit_strand"] == "Plus" else "-"
+        db_start = min(hsp["hit_from"], hsp["hit_to"]) - 1
+        query_start = min(hsp["query_from"], hsp["query_to"]) - 1
         query_end = max(hsp["query_from"], hsp["query_to"])
-        match_cigar = _match_cigar(qseq, hseq, query_start, query_end, search["query_len"])
+        query_strand = "+" if hsp["query_strand"] == "Plus" else "-"
+        alignment = Alignment(hseq=hsp["hseq"], midline=hsp["midline"], qseq=hsp["qseq"])
+        if db_strand != "+":  # ensure database is always forward
+            logger.info("Reverse-complementing match!")
+            db_strand = "+"
+            query_strand = {"+": "-", "-": "+"}[query_strand]
+            qseq = revcomp(qseq)
+            hseq = revcomp(hseq)
+            alignment = alignment.revcomp()
+        cigar = match_cigar(qseq, hseq, query_start, query_end, search["query_len"])
         return BlastMatch(
             query=os.path.basename(query),
             database=hits[0]["description"][0]["title"].split()[0],
             identity=hsp["identity"] / hsp["align_len"],
-            query_strand="+" if hsp["query_strand"] == "Plus" else "-",
+            query_strand=query_strand,
             query_start=query_start,
             query_end=query_end,
-            database_strand="+" if hsp["query_strand"] == "Plus" else "-",
+            database_strand=db_strand,
             database_start=db_start,
             database_end=max(hsp["hit_from"], hsp["hit_to"]),
-            match_cigar="".join(["".join(map(str, x)) for x in match_cigar]),
-            match_seq="".join([x for x in filter(_is_nucl, qseq)]),
-            alignment=Alignment(hseq=hsp["hseq"], midline=hsp["midline"], qseq=hsp["qseq"]),
+            match_cigar="".join(["".join(map(str, x)) for x in cigar]),
+            match_seq="".join([x for x in filter(is_nucl, qseq)]),
+            alignment=alignment,
             match_json=match_json,
         )